backend/kkomdae_flask/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126

annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
boto3==1.37.19
botocore==1.37.19
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
dotenv==0.9.9
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.0
fonttools==4.56.0
fsspec==2025.3.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
kiwisolver==1.4.8
logger==1.4
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
openai==1.70.0
opencv-python==4.11.0.86
opencv-python-headless==4.8.0.76
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.1
pydantic_core==2.33.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.4
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
supervision==0.25.1
sympy==1.13.1
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
ultralytics==8.3.99
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
Werkzeug==3.1.3
win32_setctime==1.2.0